Begin implementing automatic version field for doxygen
