+++ .Dev-App/Unity/Assets/Resources/DoxygenConfig.asset
@@ -14,14 +14,16 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Project: GUI Toolkit
   Synopsis: Toolkit with various improvements to Unity Gui
-  Version: 0.3
+  VersionSource: 0
+  m_version: 
+  OutputDirectory:
+    Path: ..\..\Documentation~\Code
   OptionalMainPage:
     Path: ..\..\README.md
   InputDirectories:
   - Path: ..\..\Runtime
   - Path: ..\..\Editor
-  OutputDirectory:
-    Path: ..\..\Documentation~\Code
-  Defines: UNITY_EDITOR
+  Defines:
+  - UNITY_EDITOR
   ExcludePatterns:
   - '*/3rdparty/*'
