+++ .Dev-App/Unity/ProjectSettings/ProjectSettings.asset
@@ -12,8 +12,8 @@ PlayerSettings:
   targetDevice: 2
   useOnDemandResources: 0
   accelerometerFrequency: 60
-  companyName: DefaultCompany
-  productName: bla
+  companyName: Arkarit
+  productName: Unity Gui Toolkit
   defaultCursor: {fileID: 0}
   cursorHotspot: {x: 0, y: 0}
   m_SplashScreenBackgroundColor: {r: 0.13725491, g: 0.12156863, b: 0.1254902, a: 1}
@@ -164,7 +164,7 @@ PlayerSettings:
   androidMinAspectRatio: 1
   applicationIdentifier:
     Android: com.UnityTestRunner.UnityTestRunner
-    Standalone: com.DefaultCompany.bla
+    Standalone: com.Arkarit.Unity-Gui-Toolkit
   buildNumber:
     Standalone: 0
     VisionOS: 0
