+++ Runtime/Code/Basic/PathField.cs
@@ -1,6 +1,4 @@
 using System;
-using System.IO;
-using UnityEngine;
 
 namespace GuiToolkit
 {
